% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconvert.R
\name{c_to_f}
\alias{c_to_f}
\title{Convert from Celsius to Fahrenheit}
\usage{
c_to_f(c_temp)
}
\arguments{
\item{c_temp}{The temperature in Celsius}
}
\value{
A numeric value
}
\description{
Convert from Celsius to Fahrenheit
}
\examples{
c_temp <- 0
c_to_f(c_temp)
}
