% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Converts Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(temp_C)
}
\arguments{
\item{temp_C}{The temperature in Celsius.}
}
\value{
The temperature in Fahrenheit.
}
\description{
This function converts input temperatures in Celsius to Fahrenheit.
}
\examples{
celsius_to_fahrenheit(0)
}
