% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c to k.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Converts Celsius to Kelvin}
\usage{
celsius_to_kelvin(temp_C)
}
\arguments{
\item{temp_C}{The temperature in Celsius.}
}
\value{
The temperature in Kelvin.
}
\description{
This function converts input temperatures in  Celsius to Kelvin.
}
\examples{
celsius_to_kelvin(32)
}
