% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconvert.R
\name{f_to_c}
\alias{f_to_c}
\title{Convert from Fahrenheit to Celsius}
\usage{
f_to_c(f_temp)
}
\arguments{
\item{f_temp}{The temperature in Fahrenheit}
}
\value{
A numeric value
}
\description{
Convert from Fahrenheit to Celsius
}
\examples{
f_temp <- 32
f_to_c(f_temp)
}
