% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahr_to_kelvin.R
\name{fahr_to_kelv}
\alias{fahr_to_kelv}
\title{Convert Fahrenheit to Kelvin}
\usage{
fahr_to_kelv(temp_f)
}
\arguments{
\item{temp}{The input temperature.}

\item{as}{a trial  of adding an argument}
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
\examples{
fahr_to_kelvin(32)
}

