% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.R
\name{fahr_to_kelvin}
\alias{fahr_to_kelvin}
\title{Converts Fahrenheit to Kelvin}
\usage{
fahr_to_kelvin(temp)
}
\arguments{
\item{temp}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
\examples{
fahr_to_kelvin(32)
}
