% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Converts Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\arguments{
\item{temp_F}{The temperature in Fahrenheit.}
}
\value{
The temperature in Celsius.
}
\description{
This function converts input temperatures in Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(32)
}
