% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Converts Fahrenheit to Kelvin}
\usage{
fahrenheit_to_kelvin(temp_F)
}
\arguments{
\item{temp_F}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin
}
\examples{
fahrenheit_to_kelvin(32)
}
