% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvin_to_celsius.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Converts Kelvin to Celsius}
\usage{
kelvin_to_celsius(temp_K)
}
\arguments{
\item{temp_K}{The temperature in Kelvin.}
}
\value{
The temperature in Celsius.
}
\description{
This function converts input temperatures in Kelvin to Celsius.
}
\examples{
kelvin_to_celsius(273.15)
}
