% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvin_to_fahrenheit.R
\name{kelvin_to_fahrenheit}
\alias{kelvin_to_fahrenheit}
\title{Converts Kelvin to Fahrenheit}
\usage{
kelvin_to_fahrenheit(temp_K)
}
\arguments{
\item{temp_K}{The temperature in Kelvin.}
}
\value{
The temperature in Fahrenheit.
}
\description{
This function converts input temperatures in Kelvin to Fahrenheit.
}
\examples{
kelvin_to_fahrenheit(273.15)
}
